automation_framework/requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-bdd==7.0.0
pytest-html==4.1.1
pytest-xdist==3.4.0
allure-pytest==2.13.2
faker==20.1.0
pyyaml==6.0.1
openpyxl==3.1.2
jsonschema==4.20.0
requests==2.31.0
selenium==4.15.2
python-dotenv==1.0.0
